+++ lib/storage_capacity.dart
@@ -11,44 +11,43 @@ class StorageCapacity {
     return version;
   }
 
-  static get getFreeSpace async {
+  static Future<String> get getFreeSpace async {
     var result = await _channel.invokeMethod('getFreeSpace');
-    return result;
+    return result.toString();
   }
 
-  static get getOccupiedSpace async {
+  static Future<String> get getOccupiedSpace async {
     var result = await _channel.invokeMethod('getOccupiedSpace');
-    return result;
+    return result.toString();
   }
 
-  static get getTotalSpace async {
+  static Future<String> get getTotalSpace async {
     var result = await _channel.invokeMethod('getTotalSpace');
-    return result;
+    return result.toString();
   }
 
-  static toMegaBytes(value) {
+  static String toMegaBytes(value) {
     return (double.parse(value) / 1000000).toStringAsFixed(2);
   }
 
-  static toGigaBytes(double value) {
+  static String toGigaBytes(double value) {
     return (value / 1000000000).toStringAsFixed(2);
   }
 
   static List<double> _capacities = [16, 32, 64, 128, 256, 512, 1024];
 
-  static searchCapacity(double totalCapacity) {
+  static String searchCapacity(double totalCapacity) {
     var i = _capacities.length ~/ 2;
     var changeCapacity = true;
 
     while (changeCapacity) {
       var difference = totalCapacity - _capacities[i];
       if (difference < 0 && difference < -4) {
-        debugPrint(_capacities[i].toString());
         i = i - 1;
       } else if ((difference < 0 && difference > -4)) {
         changeCapacity = false;
-      } else {
-        changeCapacity = false;
+      }else if(difference>0){
+        i=i+1;
       }
     }
     return _capacities[i].toString();