+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 ## 0.0.1
+initial package
 
-* TODO: Describe initial release.
+## 0.0.2
+adjusted patterns
+
+## 0.0.3
+adjusted type in return of methods
+
+## 0.0.4
+adjusted type in return of future methods